+++ tutorials/spring/00-the-basics/src/main/java/com/stormpath/tutorial/WebAppInitializer.java
@@ -15,6 +15,7 @@
  */
 package com.stormpath.tutorial;
 
+import com.stormpath.tutorial.config.WebAppConfig;
 import org.springframework.web.WebApplicationInitializer;
 import org.springframework.web.context.ContextLoaderListener;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
