--- tutorials/spring/02-spring-security-ftw/src/main/java/com/stormpath/tutorial/HelloController.java
@@ -1,56 +1,0 @@
-/*
- * Copyright 2016 Stormpath, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.stormpath.tutorial;
-
-import com.stormpath.sdk.servlet.account.AccountResolver;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.util.Assert;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import javax.servlet.http.HttpServletRequest;
-
-/**
- * @since 1.3.0
- */
-@Controller
-public class HelloController {
-
-    private HelloService helloService;
-
-    @Autowired
-    public HelloController(HelloService helloService) {
-        Assert.notNull(helloService);
-        this.helloService = helloService;
-    }
-
-    @RequestMapping("/")
-    String home(HttpServletRequest req, Model model) {
-        model.addAttribute("status", req.getParameter("status"));
-        return "home";
-    }
-
-    @RequestMapping("/restricted")
-    String restricted(HttpServletRequest req, Model model) {
-        String msg = helloService.sayHello(
-            AccountResolver.INSTANCE.getAccount(req)
-        );
-        model.addAttribute("msg", msg);
-        return "restricted";
-    }
-
-}
