--- tutorials/spring/02-spring-security-ftw/src/main/java/com/stormpath/tutorial/HelloService.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2015 Stormpath, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.stormpath.tutorial;
-
-import com.stormpath.sdk.account.Account;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Service;
-
-@Service
-public class HelloService {
-    private static final String MY_GROUP = "GROUP_HREF_HERE";
-
-    @PreAuthorize("hasRole('" + MY_GROUP + "')")
-    public String sayHello(Account account) {
-        return "Hello, " + account.getGivenName() +
-            ". You have the required permissions to access this restricted resource.";
-    }
-}
