+++ tutorials/spring/03-spring-security-refined/src/main/java/com/stormpath/tutorial/controller/ErrorController.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2016 Stormpath, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.stormpath.tutorial.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * @since 1.3.0
+ */
+@Controller
+public class ErrorController {
+
+    @RequestMapping("/403")
+    public String forbidden(Model model) {
+        Map<String, String> errors = new HashMap<>();
+        errors.put("status", "403");
+        errors.put("message", "Access is Denied");
+
+        model.addAttribute("errors", errors);
+
+        return "error";
+    }
+}
